Rename files that implement data export to match new name of XBlock.
